Changelog 0.9.5
- Updated Prime95 to version 30.19b20
- Updated y-Cruncher to version 0.8.4 Build 9538
- Updated PBO2 Tuner to the latest version
- It's now possible to catch the output from y-Cruncher, thus enabling the "auto" `runtimePerCore` setting for y-Cruncher as well. It will test all selected cores for one core and then proceed to the next one
(this is made possible by a custom wrapper .exe and .dll in the /helpers directory)
- Greatly increased the support for Intel CPUs, it now tries to detect a core architecture where some cores only support one thread (i.e. Intel's big.LITTLE) and automatically adjusts the threads it runs on these cores if two threads and `restartTestProgramForEachCore` is selected.
 Note that running with two threads and without the setting enabled may result in very long test times on the E-Cores and any test afterwards, so it's strongly recommended to either enable `restartTestProgramForEachCore`, test only with one thread, or test the P- and E-Cores separately.
- Added support for processors with more than 64 (virtual) cores for you HEDT freaks!
- Added a `testDuration` setting for y-Cruncher that allows to set the test duration in seconds for each individual test (default: 60)
- The script now also checks for WHEA errors (Windows Hardware Error Architecture) during its runtime and notices the user if such an event has happened.
It's not treated as a "real" error, as it doesn't necessarily coincide with the core that is currently being tested.
You can disable this check with the `lookForWheaErrors` setting.
- The script now also uses the Windows Event Log to store log information. This is helpful if a hard reboot occurs during testing, which can corrupt the log file. The Event Log entry should still show which core had begun testing before the hard reboot happened.
To be able to use this functionality, the script asks the user if a new "Event Source" should be added, which requires administrator privileges. This only needs to happen once, so after this "Source" has been added, no administrator privileges will be needed anymore. The script itself also doesn't need admin rights, instead it will open a new window to perform this action, and asks for elevation for this new window.
The Event Log entries can be found in the "Windows Logs"/"Applications" section of the Event Viewer, and have the source "CoreCycler".
This functionality can be controlled by the `useWindowsEventLog` setting in the `[Logging]` section.
- It's also possible to try to periodically force a "Disk Write Cache Flush", which can also help prevent log file corruption. It's not enabled by default though, to prevent possible negative performance impacts (setting `flushDiskWriteCache` in the `[Logging]` section)
- The check for the CPU utilization doesn't require Windows Performance Counters anymore, which malfunctioned far more often than I had ever imagined, and so should result in less false alarms. Instead it now uses the more readily available processor time.
If you really want to, you can re-enable the use of the Windows Performance Counters by setting `useWindowsPerformanceCountersForCpuUtilization` in the `[Debug]` section, but I really advise against it.
- Added a beep on error! (Controlled with the `beepOnError` setting)
- Added a taskbar flash on error! (Controlled with the `flashOnError` setting)
- The CPU affinity is now set to the threads of the stress test program, and not to the program/main process itself anymore.
This is required to be able to set the affinity for more than 64 CPUs, and also to fix a bug that appeared on Intel systems, where if set to use two threads, the two virtual CPUs weren't fully loaded if only the main program's affinity was set.
You won't be able to easily see the current affinity using the Task Manager anymore, but a tool like [System Informer](https://systeminformer.sourceforge.io/) can also show the current affinity for threads
- The new y-Cruncher version has updated tests and also doesn't support the `00-x86` algorithm anymore, so the new default low-load binary algorithm is `04-P4P` instead
- However, the old y-Cruncher version (which is 0.7.10) is still included, and you can use it by setting `stressTestProgram` to `YCRUNCHER_OLD`
Be aware that you will also need to adjust the `tests` setting in the config if you're switching between the versions!
- The config.default.ini is now automatically generated (and overwritten) on each script start. It doesn't have any functional use anymore and is purely there to give the user a starting point / reference for creating a custom config.ini
- The PowerShell code itself now uses the `Set-StrictMode -Version 3.0` setting. This may introduce new and fun script error messages, but hopefully I have already caught most of them!
- This caused a general code cleanup and some refactoring due to the errors I caught this way. And also eliminated a couple of bugs
- Added a LICENSE file! It just includes the text of the "CC BY-NC-SA" Creative Commons license, which the script always had anyway