Completed Python Program
Run "main.py" with python installed on your system and it is a completed and working GUI for CoreCycler. My goal is to compile all of this into a .exe file as one single file instead of having the 15 or so python files. I could use help with the compiling part of this, as I'm new to coding and keep getting errors everytime I try to compile.

Co-Authored-By: sp00n <sp.00.n@gmx.net>
Co-Authored-By: dymil <30931139+dymil@users.noreply.github.com>